Fixed Member::hashed_password to return nil on blank passwords.
The trouble was that if someone submitted the "Create Member" form without
passwords, the system would create a hash of the blank password and
then check the length of the hash, which was always long enough to pass
validation.

Now the password field isn't set if there's a blank password so that the
length validation will work as expected.

Reworked tests to provide a specific check for this situation.